requirements.txt
# LangChain Core
langchain>=0.4.0
langchain-core>=0.4.0
langchain-community>=0.4.0

# OpenAI Integration
langchain-openai
openai

# FastAPI
fastapi
uvicorn

# Vector Database
chromadb

# Database
pydantic
sqlite3

# Utilities
python-dotenv
tqdm

# Machine Learning Utilities
numpy
scikit-learn